+++ .github/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 Update <small>_ January 2024</small>
 
+- feat: migrate reportedIssues command (22/01/2024)
+- feat: docSearch command (22/01/2024)
 - fix: vatsim events max fields (20/01/2024)
 - fix: user and mod log exclude id's (18/01/2024)
 - feat: New FBW Utils and Moderation bot (08/01/2024)